+++ src-tauri/src/cmd.rs
@@ -1,9 +1,9 @@
-use crate::tasks::{sample1, sample2};
+use crate::tasks::{invoke, promise};
 use serde::Deserialize;
 
 #[derive(Debug, Deserialize)]
 #[serde(tag = "cmd")]
 pub enum Cmd {
-    Sample1(sample1::Request),
-    Sample2(sample2::Request),
+    InvokeSample(invoke::Request),
+    PromisifySample(promise::Request),
 }
